+++ app/models/user.ts
@@ -3,6 +3,7 @@ import { alias } from '@ember-decorators/object/computed';
 import { buildValidations, validator } from 'ember-cp-validations';
 import DS from 'ember-data';
 
+import Contributor from './contributor';
 import File from './file';
 import Institution from './institution';
 import Node from './node';
@@ -49,6 +50,7 @@ export default class User extends OsfModel.extend(Validations) {
     @belongsTo('region') defaultRegion!: DS.PromiseObject<Region> & Region;
 
     @hasMany('node') nodes!: DS.PromiseManyArray<Node>;
+    @hasMany('contributor') contributors!: DS.PromiseManyArray<Contributor>;
     @hasMany('registration') registrations!: DS.PromiseManyArray<Registration>;
 
     @hasMany('file') quickfiles!: DS.PromiseManyArray<File>;