+++ mirage/scenarios/default.js
@@ -1,9 +1,12 @@
-export default function(/* server */) {
-
-    /*
-      Seed your development database using your factories.
-      This data will not be loaded in your tests.
-    */
-
-    // server.createList('post', 10);
+export default function (server) {
+    const currentUser = server.create('user');
+    const firstNode = server.create('node', {});
+    server.create('contributor', { node: firstNode, users: currentUser, index: 0 });
+    const nodes = server.createList('node', 10, {}, 'withContributors');
+    server.loadFixtures('nodes');
+    for (let i = 4; i < 10; i++) {
+        server.create('contributor', { node: nodes[i], users: currentUser, index: 11 });
+    }
+    server.create('root', { currentUser });
+    server.createList('institution', 20);
 }